+++ Makefile
@@ -10,20 +10,20 @@ $(OUTPUT_DIR)/sort.o: $(OUTPUT_DIR) common/sort.h common/sort.c
 $(OUTPUT_DIR)/common.o: $(OUTPUT_DIR) common/common.h common/structures.h common/common.c
 	gcc -O2 -o $(OUTPUT_DIR)/common.o -c common/common.c
 
-$(OUTPUT_DIR)/SokobanMP_Andar.exe: brute-force/multi-thread/SokobanMP_Andar.c $(OUTPUT_DIR)/sort.o $(OUTPUT_DIR)/common.o
-	gcc -O2 -o target/SokobanMP_Andar.exe -fopenmp brute-force/multi-thread/SokobanMP_Andar.c $(OUTPUT_DIR)/*.o
+$(OUTPUT_DIR)/SokobanMP_Andar.exe: trie/SokobanMP_Andar.c $(OUTPUT_DIR)/sort.o $(OUTPUT_DIR)/common.o
+	gcc -O2 -o target/SokobanMP_Andar.exe -fopenmp trie/SokobanMP_Andar.c $(OUTPUT_DIR)/*.o
 
-$(OUTPUT_DIR)/SokobanMP_Digito.exe: brute-force/multi-thread/SokobanMP_Digito.c $(OUTPUT_DIR)/sort.o $(OUTPUT_DIR)/common.o
-	gcc -O2 -o target/SokobanMP_Digito.exe -fopenmp brute-force/multi-thread/SokobanMP_Digito.c $(OUTPUT_DIR)/*.o
+$(OUTPUT_DIR)/SokobanMP_Digito.exe: trie/SokobanMP_Digito.c $(OUTPUT_DIR)/sort.o $(OUTPUT_DIR)/common.o
+	gcc -O2 -o target/SokobanMP_Digito.exe -fopenmp trie/SokobanMP_Digito.c $(OUTPUT_DIR)/*.o
 
-$(OUTPUT_DIR)/Sokoban_HashTableList.exe: brute-force/single-thread/Sokoban_HashTableList.c $(OUTPUT_DIR)/sort.o $(OUTPUT_DIR)/common.o
-	gcc -O2 -o target/Sokoban_HashTableList.exe brute-force/single-thread/Sokoban_HashTableList.c $(OUTPUT_DIR)/*.o
+$(OUTPUT_DIR)/Sokoban_HashTableList.exe: hash-table/Sokoban_HashTableList.c $(OUTPUT_DIR)/sort.o $(OUTPUT_DIR)/common.o
+	gcc -O2 -o target/Sokoban_HashTableList.exe hash-table/Sokoban_HashTableList.c $(OUTPUT_DIR)/*.o
 
-$(OUTPUT_DIR)/Sokoban_Trie.exe: brute-force/single-thread/Sokoban_Trie.c $(OUTPUT_DIR)/sort.o $(OUTPUT_DIR)/common.o
-	gcc -O2 -o target/Sokoban_Trie.exe brute-force/single-thread/Sokoban_Trie.c $(OUTPUT_DIR)/*.o
+$(OUTPUT_DIR)/Sokoban_Trie.exe: trie/Sokoban_Trie.c $(OUTPUT_DIR)/sort.o $(OUTPUT_DIR)/common.o
+	gcc -O2 -o target/Sokoban_Trie.exe trie/Sokoban_Trie.c $(OUTPUT_DIR)/*.o
 
-$(OUTPUT_DIR)/Sokoban_Heuristic.exe: heuristic/sokoban_heuristic.c $(OUTPUT_DIR)/sort.o $(OUTPUT_DIR)/common.o
-	gcc -O2 -o target/Sokoban_Heuristic.exe heuristic/sokoban_heuristic.c $(OUTPUT_DIR)/*.o
+$(OUTPUT_DIR)/Sokoban_Heuristic.exe: hash-table/Sokoban_Heuristic.c $(OUTPUT_DIR)/sort.o $(OUTPUT_DIR)/common.o
+	gcc -O2 -o target/Sokoban_Heuristic.exe hash-table/Sokoban_Heuristic.c $(OUTPUT_DIR)/*.o
 
 $(OUTPUT_DIR):
 	mkdir -p $(OUTPUT_DIR)
